requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
groq==0.4.1
PyMuPDF==1.23.7
python-docx==1.0.1
pdfplumber==0.10.2
fpdf==1.7.2 